requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
exceptiongroup==1.3.1
Flask==3.1.2
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.1.53
langgraph==0.0.40
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.1.147
MarkupSafe==3.0.3
matplotlib==3.7.2
mysql-connector-python==9.4.0
noise==1.2.2
numpy==1.24.3
openai==2.8.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==23.2
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.5.2
PyJWT==2.10.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.4
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
